atria_metrics/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atria_metrics"
version = "0.0.0"                                                                                                    
description = "Atria Metrics: A collection of machine learning metrics for evaluating models in the Atria ecosystem."
readme = "README.md"
requires-python = "==3.11.*"
license = { file = "LICENSE.txt" }
authors = [{ name = "Saifullah", email = "saifullah.saifullah@dfki.de" }]
dependencies = [
    "anls>=0.0.2",
    "atria-registry",
    "atria-types",
    "pytorch-ignite>=0.5.2",
    "seqeval",
]

[tool.hatch.metadata]
allow-direct-references = true
test = ["coverage", "pytest"]

[tool.coverage.report]
skip_covered = true
show_missing = true
precision = 2
exclude_lines = [
    'def __repr__',
    'pragma= no cover',
    'raise NotImplementedError',
    'if TYPE_CHECKING=',
    'if typing.TYPE_CHECKING=',
    '@overload',
    '@typing.overload',
    '\(Protocol\)=$',
    'typing.assert_never',
    'assert_never',
    'if __name__ == "__main__":',
]

## Linter and formatter
[tool.ruff]
# cover and extend the default config in https=//docs.astral.sh/ruff/configuration/
extend-exclude = ["alembic"]
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long, handled by black
    "B008",   # do not perform function calls in argument defaults
    "W191",   # indentation contains tabs
    "B904",   # Allow raising exceptions without from e, for HTTPException
    "COM819", # Trailing comma prohibited
    "D100",   # Missing docstring in public module(file)
    "D104",   # Missing docstring in public package
    "D203",   # 1 blank line required before class docstring
    "E201",   # Whitespace after '('
    "E202",   # Whitespace before ')'
    "E203",   # Whitespace before '='
    "E221",   # Multiple spaces before operator
    "E241",   # Multiple spaces after ','
    "E251",   # Unexpected spaces around keyword / parameter equals
    "W291",   # Trailing whitespace
    "W293",   # Blank line contains whitespace
]

isort = { combine-as-imports = true,  split-on-trailing-comma = false }

# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

[tool.ruff.format]
docstring-code-format = true
skip-magic-trailing-comma = true

[project.urls]
"Homepage" = "https://github.com/saifullah3396/atria/"
"Bug Reports" = "https://github.com/saifullah3396/atria/"
"Source" = "https://github.com/saifullah3396/atria/"

[tool.pytest.ini_options]
pythongit = [".", "src", "tests"]
# Set additional command line options for pytest
# Ref: https://docs.pytest.org/en/stable/reference/reference.html#command-line-flags
addopts = "-rXs --strict-config --strict-markers --tb=short"
xfail_strict = true         # Treat tests that are marked as xfail but pass as test failures
# filterwarnings = ["error"]  # Treat all warnings as errors

[tool.uv]
managed = true

[tool.uv.sources]
atria-types = { path = "../atria_types", editable = true }
atria-registry = { path = "../atria_registry", editable = true }

[tool.mypy]
disable_error_code = ["import-untyped", "import-not-found"]
